Basic division to dev and prod configs add favicon
