+++ webpack/webpack.config.dev.js
@@ -1,13 +1,10 @@
-const ExtractTextPlugin = require("extract-text-webpack-plugin");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
+const configCommon = require("./webpack.config.common");
+const merge = require("webpack-merge");
 const webpack = require("webpack");
 
 const paths = require("./paths");
-const loadAppEnv = require("./loadAppEnv");
 
-const applicationEnv = loadAppEnv(process.env);
-
-module.exports = {
+module.exports = merge(configCommon, {
   devServer: {
     contentBase: paths.dist,
     host: "localhost",
@@ -27,16 +24,8 @@ module.exports = {
     "react-hot-loader/patch",
     "webpack-dev-server/client?http://localhost:9090",
     "webpack/hot/only-dev-server",
-    "./app/index.tsx",
   ],
-  resolve: {
-    extensions: [".ts", ".tsx", ".js"],
-  },
-  output: {
-    path: paths.dist,
-    filename: "[name].js",
-    publicPath: "/",
-  },
+  plugins: [new webpack.NamedModulesPlugin(), new webpack.HotModuleReplacementPlugin()],
   module: {
     rules: [
       {
@@ -58,73 +47,6 @@ module.exports = {
         ],
         include: paths.app,
       },
-      {
-        oneOf: [
-          {
-            test: /\.module.scss$/,
-            use: [
-              {
-                loader: "style-loader",
-              },
-              {
-                loader: "css-loader",
-                options: {
-                  importLoaders: 1,
-                  modules: true,
-                  localIdentName: "[name]__[local]___[hash:base64:5]",
-                  camelCase: "dashesOnly",
-                },
-              },
-              { loader: "sass-loader" },
-            ],
-          },
-          {
-            test: /\.scss$/,
-            use: [
-              {
-                loader: "style-loader",
-              },
-              {
-                loader: "css-loader",
-                options: {
-                  importLoaders: 1,
-                  modules: false,
-                  localIdentName: "[name]__[local]___[hash:base64:5]",
-                  camelCase: "dashesOnly",
-                },
-              },
-              { loader: "sass-loader" },
-            ],
-          },
-          {
-            test: /\.(jpg|png|svg)$/,
-            loader: "url-loader",
-            options: {
-              limit: 25000,
-              publicPath: "/",
-            },
-          },
-          {
-            test: /\.(woff2|woff|ttf|eot|otf)$/,
-            loader: "file-loader",
-            options: {
-              name: "fonts/[hash].[ext]",
-            },
-          },
-        ],
-      },
     ],
   },
-  plugins: [
-    new ExtractTextPlugin("styles.css"),
-    new HtmlWebpackPlugin({
-      inject: true,
-      template: paths.appHtml,
-    }),
-    new webpack.NamedModulesPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.DefinePlugin({
-      "process.env": applicationEnv,
-    }),
-  ],
-};
+});
