+++ webpack/webpack.config.prod.js
@@ -1,0 +1,35 @@
+const ExtractTextPlugin = require("extract-text-webpack-plugin");
+const CleanWebpackPlugin = require("clean-webpack-plugin");
+const merge = require("webpack-merge");
+
+const configCommon = require("./webpack.config.common");
+const paths = require("./paths");
+
+module.exports = merge(configCommon, {
+  output: {
+    filename: "[name].min.js",
+  },
+  plugins: [
+    new CleanWebpackPlugin(paths.dist, {
+      root: paths.root,
+    }),
+    new ExtractTextPlugin("styles.css"), // TODO: its not working
+  ],
+  module: {
+    rules: [
+      {
+        test: /\.(tsx?)$/,
+        use: [
+          {
+            loader: "awesome-typescript-loader",
+            options: {
+              configFileName: "./tsconfig.json",
+              useCache: false,
+            },
+          },
+        ],
+        include: paths.app,
+      },
+    ],
+  },
+});
